feat(#25): code cleanup and module updates
